requirements.txt
flask==2.0.1
requests==2.31.0
python-dotenv==1.0.0
gTTS==2.3.2
SpeechRecognition==3.10.0
pydub==0.25.1
pyaudio==0.2.13
playsound==1.2.2 